Add reboot allowance toggle
NOTE: This change changes the default behaviour of the Role

Added a variable 'update_upgrade_reboot_allow_reboot'.
This variable defaults to 'False'.
When, after an update, the server needs a reboot, we no longer
perform the reboot action, but will display a warning message.
It is then up to the user to perform the reboot at his convenience.
If the variable 'update_upgrade_reboot_allow_reboot' is set to 'True',
we will automatically reboot the host.